Fix timeout and Windows AMI in aws_ssm (#235)
Windows AMIs get replaced every few months, don't be specific with image name

Task timeout feature in ansible-base (#69284) resulted in timeout key not being
properly passed to plugins. This was fixed in (#71722). The intent for the
timeout option in this plugin is to not use Ansible's default timeout however,
but to defer to the inventory var:
https://github.com/ansible/ansible/pull/49652/commits/44f1427507d588d2748c7db34e590234de661564
Rename this internal value from `timeout` to `plugin_timeout`.