Improve state and config changes in ec2_instance
DO NOT MERGE!!!
Fixes: community.aws/issues/16
This patch ensures that when changing state and/or supported parameters for
an instance, changes are applied regardless of the starting or desired state
of the instance.

This patch depends on Python3 and SHOULD NOT be merged until this module is
migrated to the amazon.aws collection and collection version 2.x.