Fix idempotency and logic to update existing aggregator
`describe_configuration_aggregators` method returns output similar
to the following:

```
{
    "ConfigurationAggregators": [
        {
            "ConfigurationAggregatorName": "test-name",
            "ConfigurationAggregatorArn": "arn:aws:config:us-east-1:123412341234:config-aggregator/config-aggregator-xbxbvyq5",
            "OrganizationAggregationSource": {
                "RoleArn": "arn:aws:iam::123412341234:role/my-role",
                "AllAwsRegions": true
            },
            "CreationTime": 1619030767.047,
            "LastUpdatedTime": 1626463216.998
        }
    ]
}
```

As a result, lines 134-136 fail:

```
    del current_params['ConfigurationAggregatorArn']
    del current_params['CreationTime']
    del current_params['LastUpdatedTime']
```
as they try to delete attributes from the `current_params` as opposed
to `current_params['ConfigurationAggregators'][0]`.

The error message is:
```
KeyError: 'ConfigurationAggregators'
```

Additionally, if no `account_sources` attribute is specified, the module
fails idempotency check, because in that case `AccountAggregationSources`
attribute is present in `params`, but not in `current_params`.