glue_connection - Avoid converting connection_parameter keys to lowercase (#518)
glue_connection - Avoid converting connection_parameter keys to lowercase

SUMMARY

This is a follow to my own PR #503.
This is a cosmetic change that prevents converting keys in connection_parameters dict to lowercase.

ISSUE TYPE


Bugfix Pull Request

COMPONENT NAME

aws_glue_connection
ADDITIONAL INFORMATION



As an example, this:
- community.aws.glue_connection:
    name: test-connection
    connection_parameters:
      JDBC_ENFORCE_SSL: "false"
    ...

is a valid value, while this:
- community.aws.glue_connection:
    name: test-connection
    connection_parameters:
      jdbc_enforce_ssl: "false"
    ...

is not.
This PR simply aligns the module output to the expected input.

Reviewed-by: Mark Chappell <None>