Remove new ec2_snapshot_copy tests for now
The new tests seem to be failing due to a permissions issue with the
service account used to run them. Since they were marked as optional
in the MR I'm going to remove them but leave them in the history[1] so
the work's not lost in case someone wants to use them.

[1] The tests are still on aec7b2fd397337fced67d9b4846fbeed74db61cc