Change name of server certs in integration tests
There is a mismatch in the terminator repo for the permissions on iam
server certs and the terminator class for server certs. This has led to
the CI account reaching the quota for server certs, as some of the older
certs are not getting cleaned up.

I don't know the reasoning behind limiting the terminator class to only
certs prefixed with ansible-test. The safest option here seems to be to
change the permissions to only allow creating certs with this prefix.
Once this update is merged I'll change the permissions on the terminator
repo and manually clear out the older certs.